third_party/blink/web_tests/virtual/stable/http/tests/worklet/webexposed/global-interface-listing-paint-worklet-expected.txt

CONSOLE MESSAGE: This test logs exposed APIs once from each PaintWorkletGlobalScope
CONSOLE MESSAGE: [INTERFACES]
CONSOLE MESSAGE: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSImageValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSKeywordValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter value
CONSOLE MESSAGE: interface CSSMathClamp : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter lower
CONSOLE MESSAGE:     getter upper
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathInvert : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathMax : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter values
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathMin : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter values
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathNegate : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathProduct : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter values
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathSum : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter values
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathValue : CSSNumericValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter operator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMatrixComponent : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter matrix
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter matrix
CONSOLE MESSAGE: interface CSSNumericArray
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter length
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface CSSNumericValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method add
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method div
CONSOLE MESSAGE:     method equals
CONSOLE MESSAGE:     method max
CONSOLE MESSAGE:     method min
CONSOLE MESSAGE:     method mul
CONSOLE MESSAGE:     method sub
CONSOLE MESSAGE:     method to
CONSOLE MESSAGE:     method toSum
CONSOLE MESSAGE:     method type
CONSOLE MESSAGE: interface CSSPerspective : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter length
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter length
CONSOLE MESSAGE: interface CSSPositionValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter x
CONSOLE MESSAGE:     getter y
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter x
CONSOLE MESSAGE:     setter y
CONSOLE MESSAGE: interface CSSRotate : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter angle
CONSOLE MESSAGE:     getter x
CONSOLE MESSAGE:     getter y
CONSOLE MESSAGE:     getter z
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter angle
CONSOLE MESSAGE:     setter x
CONSOLE MESSAGE:     setter y
CONSOLE MESSAGE:     setter z
CONSOLE MESSAGE: interface CSSScale : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter x
CONSOLE MESSAGE:     getter y
CONSOLE MESSAGE:     getter z
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter x
CONSOLE MESSAGE:     setter y
CONSOLE MESSAGE:     setter z
CONSOLE MESSAGE: interface CSSSkew : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter ax
CONSOLE MESSAGE:     getter ay
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter ax
CONSOLE MESSAGE:     setter ay
CONSOLE MESSAGE: interface CSSSkewX : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter ax
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter ax
CONSOLE MESSAGE: interface CSSSkewY : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter ay
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter ay
CONSOLE MESSAGE: interface CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method toString
CONSOLE MESSAGE: interface CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter is2D
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method toMatrix
CONSOLE MESSAGE:     method toString
CONSOLE MESSAGE:     setter is2D
CONSOLE MESSAGE: interface CSSTransformValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter is2D
CONSOLE MESSAGE:     getter length
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method toMatrix
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface CSSTranslate : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter x
CONSOLE MESSAGE:     getter y
CONSOLE MESSAGE:     getter z
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter x
CONSOLE MESSAGE:     setter y
CONSOLE MESSAGE:     setter z
CONSOLE MESSAGE: interface CSSUnitValue : CSSNumericValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter unit
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter value
CONSOLE MESSAGE: interface CSSUnparsedValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter length
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface CSSVariableReferenceValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter fallback
CONSOLE MESSAGE:     getter variable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter variable
CONSOLE MESSAGE: interface CountQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface PaintRenderingContext2D
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter fillStyle
CONSOLE MESSAGE:     getter filter
CONSOLE MESSAGE:     getter globalAlpha
CONSOLE MESSAGE:     getter globalCompositeOperation
CONSOLE MESSAGE:     getter imageSmoothingEnabled
CONSOLE MESSAGE:     getter lineCap
CONSOLE MESSAGE:     getter lineDashOffset
CONSOLE MESSAGE:     getter lineJoin
CONSOLE MESSAGE:     getter lineWidth
CONSOLE MESSAGE:     getter miterLimit
CONSOLE MESSAGE:     getter shadowBlur
CONSOLE MESSAGE:     getter shadowColor
CONSOLE MESSAGE:     getter shadowOffsetX
CONSOLE MESSAGE:     getter shadowOffsetY
CONSOLE MESSAGE:     getter strokeStyle
CONSOLE MESSAGE:     method arc
CONSOLE MESSAGE:     method arcTo
CONSOLE MESSAGE:     method beginPath
CONSOLE MESSAGE:     method bezierCurveTo
CONSOLE MESSAGE:     method clearRect
CONSOLE MESSAGE:     method clip
CONSOLE MESSAGE:     method closePath
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method createConicGradient
CONSOLE MESSAGE:     method createLinearGradient
CONSOLE MESSAGE:     method createPattern
CONSOLE MESSAGE:     method createRadialGradient
CONSOLE MESSAGE:     method drawImage
CONSOLE MESSAGE:     method ellipse
CONSOLE MESSAGE:     method fill
CONSOLE MESSAGE:     method fillRect
CONSOLE MESSAGE:     method getLineDash
CONSOLE MESSAGE:     method getTransform
CONSOLE MESSAGE:     method isPointInPath
CONSOLE MESSAGE:     method isPointInStroke
CONSOLE MESSAGE:     method lineTo
CONSOLE MESSAGE:     method moveTo
CONSOLE MESSAGE:     method quadraticCurveTo
CONSOLE MESSAGE:     method rect
CONSOLE MESSAGE:     method reset
CONSOLE MESSAGE:     method resetTransform
CONSOLE MESSAGE:     method restore
CONSOLE MESSAGE:     method rotate
CONSOLE MESSAGE:     method roundRect
CONSOLE MESSAGE:     method save
CONSOLE MESSAGE:     method scale
CONSOLE MESSAGE:     method setLineDash
CONSOLE MESSAGE:     method setTransform
CONSOLE MESSAGE:     method stroke
CONSOLE MESSAGE:     method strokeRect
CONSOLE MESSAGE:     method transform
CONSOLE MESSAGE:     method translate
CONSOLE MESSAGE:     setter fillStyle
CONSOLE MESSAGE:     setter filter
CONSOLE MESSAGE:     setter globalAlpha
CONSOLE MESSAGE:     setter globalCompositeOperation
CONSOLE MESSAGE:     setter imageSmoothingEnabled
CONSOLE MESSAGE:     setter lineCap
CONSOLE MESSAGE:     setter lineDashOffset
CONSOLE MESSAGE:     setter lineJoin
CONSOLE MESSAGE:     setter lineWidth
CONSOLE MESSAGE:     setter miterLimit
CONSOLE MESSAGE:     setter shadowBlur
CONSOLE MESSAGE:     setter shadowColor
CONSOLE MESSAGE:     setter shadowOffsetX
CONSOLE MESSAGE:     setter shadowOffsetY
CONSOLE MESSAGE:     setter strokeStyle
CONSOLE MESSAGE: interface PaintSize
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter height
CONSOLE MESSAGE:     getter width
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface PaintWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface Path2D
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method addPath
CONSOLE MESSAGE:     method arc
CONSOLE MESSAGE:     method arcTo
CONSOLE MESSAGE:     method bezierCurveTo
CONSOLE MESSAGE:     method closePath
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method ellipse
CONSOLE MESSAGE:     method lineTo
CONSOLE MESSAGE:     method moveTo
CONSOLE MESSAGE:     method quadraticCurveTo
CONSOLE MESSAGE:     method rect
CONSOLE MESSAGE:     method roundRect
CONSOLE MESSAGE: interface ReadableByteStreamController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter byobRequest
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method @@asyncIterator
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getReader
CONSOLE MESSAGE:     method pipeThrough
CONSOLE MESSAGE:     method pipeTo
CONSOLE MESSAGE:     method tee
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface ReadableStreamBYOBReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface ReadableStreamBYOBRequest
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter view
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method respond
CONSOLE MESSAGE:     method respondWithNewView
CONSOLE MESSAGE: interface ReadableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStreamDefaultReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface StylePropertyMapReadOnly
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method get
CONSOLE MESSAGE:     method getAll
CONSOLE MESSAGE:     method has
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface TransformStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface TransformStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method terminate
CONSOLE MESSAGE: interface WorkletGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface WritableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getWriter
CONSOLE MESSAGE: interface WritableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter signal
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface WritableStreamDefaultWriter
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     getter ready
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE:     method write
CONSOLE MESSAGE: [NAMESPACES]
CONSOLE MESSAGE: namespace console
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method assert
CONSOLE MESSAGE:     method clear
CONSOLE MESSAGE:     method context
CONSOLE MESSAGE:     method count
CONSOLE MESSAGE:     method countReset
CONSOLE MESSAGE:     method createTask
CONSOLE MESSAGE:     method debug
CONSOLE MESSAGE:     method dir
CONSOLE MESSAGE:     method dirxml
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method group
CONSOLE MESSAGE:     method groupCollapsed
CONSOLE MESSAGE:     method groupEnd
CONSOLE MESSAGE:     method info
CONSOLE MESSAGE:     method log
CONSOLE MESSAGE:     method profile
CONSOLE MESSAGE:     method profileEnd
CONSOLE MESSAGE:     method table
CONSOLE MESSAGE:     method time
CONSOLE MESSAGE:     method timeEnd
CONSOLE MESSAGE:     method timeLog
CONSOLE MESSAGE:     method timeStamp
CONSOLE MESSAGE:     method trace
CONSOLE MESSAGE:     method warn
CONSOLE MESSAGE: [GLOBAL OBJECT]
CONSOLE MESSAGE:     attribute globalThis
CONSOLE MESSAGE:     getter devicePixelRatio
CONSOLE MESSAGE:     method gc
CONSOLE MESSAGE:     method registerPaint
CONSOLE MESSAGE: [INTERFACES]
CONSOLE MESSAGE: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSImageValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSKeywordValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter value
CONSOLE MESSAGE: interface CSSMathClamp : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter lower
CONSOLE MESSAGE:     getter upper
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathInvert : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathMax : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter values
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathMin : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter values
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathNegate : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathProduct : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter values
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathSum : CSSMathValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter values
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMathValue : CSSNumericValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter operator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CSSMatrixComponent : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter matrix
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter matrix
CONSOLE MESSAGE: interface CSSNumericArray
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter length
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface CSSNumericValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method add
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method div
CONSOLE MESSAGE:     method equals
CONSOLE MESSAGE:     method max
CONSOLE MESSAGE:     method min
CONSOLE MESSAGE:     method mul
CONSOLE MESSAGE:     method sub
CONSOLE MESSAGE:     method to
CONSOLE MESSAGE:     method toSum
CONSOLE MESSAGE:     method type
CONSOLE MESSAGE: interface CSSPerspective : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter length
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter length
CONSOLE MESSAGE: interface CSSPositionValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter x
CONSOLE MESSAGE:     getter y
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter x
CONSOLE MESSAGE:     setter y
CONSOLE MESSAGE: interface CSSRotate : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter angle
CONSOLE MESSAGE:     getter x
CONSOLE MESSAGE:     getter y
CONSOLE MESSAGE:     getter z
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter angle
CONSOLE MESSAGE:     setter x
CONSOLE MESSAGE:     setter y
CONSOLE MESSAGE:     setter z
CONSOLE MESSAGE: interface CSSScale : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter x
CONSOLE MESSAGE:     getter y
CONSOLE MESSAGE:     getter z
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter x
CONSOLE MESSAGE:     setter y
CONSOLE MESSAGE:     setter z
CONSOLE MESSAGE: interface CSSSkew : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter ax
CONSOLE MESSAGE:     getter ay
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter ax
CONSOLE MESSAGE:     setter ay
CONSOLE MESSAGE: interface CSSSkewX : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter ax
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter ax
CONSOLE MESSAGE: interface CSSSkewY : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter ay
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter ay
CONSOLE MESSAGE: interface CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method toString
CONSOLE MESSAGE: interface CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter is2D
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method toMatrix
CONSOLE MESSAGE:     method toString
CONSOLE MESSAGE:     setter is2D
CONSOLE MESSAGE: interface CSSTransformValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter is2D
CONSOLE MESSAGE:     getter length
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method toMatrix
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface CSSTranslate : CSSTransformComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter x
CONSOLE MESSAGE:     getter y
CONSOLE MESSAGE:     getter z
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter x
CONSOLE MESSAGE:     setter y
CONSOLE MESSAGE:     setter z
CONSOLE MESSAGE: interface CSSUnitValue : CSSNumericValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter unit
CONSOLE MESSAGE:     getter value
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter value
CONSOLE MESSAGE: interface CSSUnparsedValue : CSSStyleValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter length
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface CSSVariableReferenceValue
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter fallback
CONSOLE MESSAGE:     getter variable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     setter variable
CONSOLE MESSAGE: interface CountQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface PaintRenderingContext2D
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter fillStyle
CONSOLE MESSAGE:     getter filter
CONSOLE MESSAGE:     getter globalAlpha
CONSOLE MESSAGE:     getter globalCompositeOperation
CONSOLE MESSAGE:     getter imageSmoothingEnabled
CONSOLE MESSAGE:     getter lineCap
CONSOLE MESSAGE:     getter lineDashOffset
CONSOLE MESSAGE:     getter lineJoin
CONSOLE MESSAGE:     getter lineWidth
CONSOLE MESSAGE:     getter miterLimit
CONSOLE MESSAGE:     getter shadowBlur
CONSOLE MESSAGE:     getter shadowColor
CONSOLE MESSAGE:     getter shadowOffsetX
CONSOLE MESSAGE:     getter shadowOffsetY
CONSOLE MESSAGE:     getter strokeStyle
CONSOLE MESSAGE:     method arc
CONSOLE MESSAGE:     method arcTo
CONSOLE MESSAGE:     method beginPath
CONSOLE MESSAGE:     method bezierCurveTo
CONSOLE MESSAGE:     method clearRect
CONSOLE MESSAGE:     method clip
CONSOLE MESSAGE:     method closePath
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method createConicGradient
CONSOLE MESSAGE:     method createLinearGradient
CONSOLE MESSAGE:     method createPattern
CONSOLE MESSAGE:     method createRadialGradient
CONSOLE MESSAGE:     method drawImage
CONSOLE MESSAGE:     method ellipse
CONSOLE MESSAGE:     method fill
CONSOLE MESSAGE:     method fillRect
CONSOLE MESSAGE:     method getLineDash
CONSOLE MESSAGE:     method getTransform
CONSOLE MESSAGE:     method isPointInPath
CONSOLE MESSAGE:     method isPointInStroke
CONSOLE MESSAGE:     method lineTo
CONSOLE MESSAGE:     method moveTo
CONSOLE MESSAGE:     method quadraticCurveTo
CONSOLE MESSAGE:     method rect
CONSOLE MESSAGE:     method reset
CONSOLE MESSAGE:     method resetTransform
CONSOLE MESSAGE:     method restore
CONSOLE MESSAGE:     method rotate
CONSOLE MESSAGE:     method roundRect
CONSOLE MESSAGE:     method save
CONSOLE MESSAGE:     method scale
CONSOLE MESSAGE:     method setLineDash
CONSOLE MESSAGE:     method setTransform
CONSOLE MESSAGE:     method stroke
CONSOLE MESSAGE:     method strokeRect
CONSOLE MESSAGE:     method transform
CONSOLE MESSAGE:     method translate
CONSOLE MESSAGE:     setter fillStyle
CONSOLE MESSAGE:     setter filter
CONSOLE MESSAGE:     setter globalAlpha
CONSOLE MESSAGE:     setter globalCompositeOperation
CONSOLE MESSAGE:     setter imageSmoothingEnabled
CONSOLE MESSAGE:     setter lineCap
CONSOLE MESSAGE:     setter lineDashOffset
CONSOLE MESSAGE:     setter lineJoin
CONSOLE MESSAGE:     setter lineWidth
CONSOLE MESSAGE:     setter miterLimit
CONSOLE MESSAGE:     setter shadowBlur
CONSOLE MESSAGE:     setter shadowColor
CONSOLE MESSAGE:     setter shadowOffsetX
CONSOLE MESSAGE:     setter shadowOffsetY
CONSOLE MESSAGE:     setter strokeStyle
CONSOLE MESSAGE: interface PaintSize
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter height
CONSOLE MESSAGE:     getter width
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface PaintWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface Path2D
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method addPath
CONSOLE MESSAGE:     method arc
CONSOLE MESSAGE:     method arcTo
CONSOLE MESSAGE:     method bezierCurveTo
CONSOLE MESSAGE:     method closePath
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method ellipse
CONSOLE MESSAGE:     method lineTo
CONSOLE MESSAGE:     method moveTo
CONSOLE MESSAGE:     method quadraticCurveTo
CONSOLE MESSAGE:     method rect
CONSOLE MESSAGE:     method roundRect
CONSOLE MESSAGE: interface ReadableByteStreamController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter byobRequest
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method @@asyncIterator
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getReader
CONSOLE MESSAGE:     method pipeThrough
CONSOLE MESSAGE:     method pipeTo
CONSOLE MESSAGE:     method tee
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface ReadableStreamBYOBReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface ReadableStreamBYOBRequest
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter view
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method respond
CONSOLE MESSAGE:     method respondWithNewView
CONSOLE MESSAGE: interface ReadableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStreamDefaultReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface StylePropertyMapReadOnly
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method get
CONSOLE MESSAGE:     method getAll
CONSOLE MESSAGE:     method has
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface TransformStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface TransformStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method terminate
CONSOLE MESSAGE: interface WorkletGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface WritableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getWriter
CONSOLE MESSAGE: interface WritableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter signal
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface WritableStreamDefaultWriter
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     getter ready
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE:     method write
CONSOLE MESSAGE: [NAMESPACES]
CONSOLE MESSAGE: namespace console
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method assert
CONSOLE MESSAGE:     method clear
CONSOLE MESSAGE:     method context
CONSOLE MESSAGE:     method count
CONSOLE MESSAGE:     method countReset
CONSOLE MESSAGE:     method createTask
CONSOLE MESSAGE:     method debug
CONSOLE MESSAGE:     method dir
CONSOLE MESSAGE:     method dirxml
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method group
CONSOLE MESSAGE:     method groupCollapsed
CONSOLE MESSAGE:     method groupEnd
CONSOLE MESSAGE:     method info
CONSOLE MESSAGE:     method log
CONSOLE MESSAGE:     method profile
CONSOLE MESSAGE:     method profileEnd
CONSOLE MESSAGE:     method table
CONSOLE MESSAGE:     method time
CONSOLE MESSAGE:     method timeEnd
CONSOLE MESSAGE:     method timeLog
CONSOLE MESSAGE:     method timeStamp
CONSOLE MESSAGE:     method trace
CONSOLE MESSAGE:     method warn
CONSOLE MESSAGE: [GLOBAL OBJECT]
CONSOLE MESSAGE:     attribute globalThis
CONSOLE MESSAGE:     getter devicePixelRatio
CONSOLE MESSAGE:     method gc
CONSOLE MESSAGE:     method registerPaint
